pyproject.toml
[project]
name = "pystruct"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "mostafa2k", email = "m2kappswindows8.1@gmail.com"},
]
dependencies = ["pyyaml>=6.0.2"]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
